# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rede Akiba API REST",
        "description": "Esta documentação descreve as todas as operações disponíveis para as aplicações da Rede Akiba.",
        "contact": {
            "email": "contato@redeakiba.com.br"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/repositorio-de-arquivos": {
            "get": {
                "tags": [
                    "Repositório de Arquivos"
                ],
                "summary": "Lista todos os arquivos cadastrados",
                "description": "Este endpoint retorna todos os arquivos cadastrados no sistema",
                "operationId": "cac14121464622393812d8895204f15f",
                "responses": {
                    "200": {
                        "description": "Arquivos encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FilesRepositoryResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nenhum arquivo encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Nenhuma arquivo encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro de processamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ocorreu um erro de processamento"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Repositório de Arquivos"
                ],
                "summary": "Cadastra um novo arquivo",
                "description": "Este endpoint realiza o cadastro de um novo arquivo no sistema",
                "operationId": "8d94630512f6e44be4033f72047786f5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FilesRepositoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Arquivo cadastrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FilesRepositoryResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ocorreu um erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ocorreu um erro de validação"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro de processamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ocorreu um erro de processamento"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/repositorio-de-arquivos/{id}": {
            "get": {
                "tags": [
                    "Repositório de Arquivos"
                ],
                "summary": "Retorna um arquivo específico",
                "description": "Este endpoint retorna um arquivo específico cadastrado no sistema",
                "operationId": "a3c57b2188425ce6b9fc075497e42913",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Arquivo: Retorna um arquivo específico baseado no id fornecido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Arquivo encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FilesRepositoryResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Arquivo não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Arquivo não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro de processamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ocorreu um erro de processamento"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Repositório de Arquivos"
                ],
                "summary": "Remove um arquivo específico",
                "description": "Este endpoint remove um arquivo específico cadastrado no sistema",
                "operationId": "495f653238e3bd97d8da1627dc885ea3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Arquivo: Remove um arquivo específico baseado no id fornecido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Arquivo removido com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Arquivo removido com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Arquivo não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Arquivo não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro de processamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ocorreu um erro de processamento"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Repositório de Arquivos"
                ],
                "summary": "Atualiza um arquivo específico",
                "description": "Este endpoint atualiza um arquivo específico cadastrado no sistema",
                "operationId": "cb17c80533602e36934cd73cd5c0cfda",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Arquivo: Atualiza um arquivo específico baseado no id fornecido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FilesRepositoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Arquivo atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FilesRepositoryResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Arquivo não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Arquivo não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro de processamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ocorreu um erro de processamento"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "FilesRepositoryResponse": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "category": {
                        "type": "string"
                    },
                    "icon": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "FilesRepositoryRequest": {
                "properties": {
                    "category": {
                        "type": "string"
                    },
                    "icon": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}